feat: update-P page separate
TODO:
need to fix
- delete button [modal different + normal different] / delete on the main page
- testing (add + delete)
- heading name change
- suspense add
  Admin -> Details page modal change + modify
             - > Order Table + Next Ui Data Table / Shadcn
Extra TODO :
   Client Side -
     1. Product List Page (try add to cart button on card), Pagination , Filter , Sort  , Navbar, Banner
     2. Details page
     3. (⚠️)  User Page +  Add to Cart , User Setting Page , Order Details Page , Cart Page , Order with Pay page

UI fix -> Multiple Filter + Skeleton + Loader fix

Backend - Order with payment , authentication(reset pass) + otp, mail generate , file upload , route rearrange with permission

Extra : add to cart (if user not login)
