+++ client_admin_panel/src/app/(admin-routes)/(product-t1)/(operation)/product-t1-update-form/layout.tsx
@@ -1,0 +1,3 @@
+export default function Layout({ children }: { children: React.ReactNode }) {
+  return <>{children}</>;
+}
