+++ client_admin_panel/src/app/(admin-routes)/(product-t1)/(operation)/product-t1-update-form/update/[id]/page.tsx
@@ -1,17 +1,16 @@
-"use client";
-import "@/styles/globals.css";
 import React from "react";
-
 import IfUserThenIfAdmin from "@/providers/security/if-user-then-if-admin";
 import PageMergin from "@/wrapper/page-mergin";
-import UiTabChange from "@/components/product-t1/products/list/ui-tab-change";
+import ProductUpdateLayout from "@/components/product-t1/operation/product/update/product-update-layout";
 
-export default function Home() {
+const Page = () => {
   return (
     <IfUserThenIfAdmin>
       <PageMergin>
-        <UiTabChange />
+        <ProductUpdateLayout />
       </PageMergin>
     </IfUserThenIfAdmin>
   );
-}
+};
+
+export default Page;
