--- client_admin_panel/src/app/(admin-routes)/(product-t1)/product-t1-add-form/page.tsx
@@ -1,16 +1,0 @@
-import React from "react";
-import IfUserThenIfAdmin from "@/providers/security/if-user-then-if-admin";
-import ProductAddLayout from "@/components/product-t1/create/product/product-add-layout";
-import PageMergin from "@/wrapper/page-mergin";
-
-const Page = () => {
-  return (
-    <IfUserThenIfAdmin>
-      <PageMergin>
-        <ProductAddLayout />
-      </PageMergin>
-    </IfUserThenIfAdmin>
-  );
-};
-
-export default Page;
