--- client_admin_panel/src/app/(admin-routes)/(product-t1)/product-t1-edit-form/[id]/page.tsx
@@ -1,16 +1,0 @@
-import React from "react";
-import ProductForm from "@/components/product-t1/update/update-product";
-import IfUserThenIfAdmin from "@/providers/security/if-user-then-if-admin";
-import PageMergin from "@/wrapper/page-mergin";
-
-const Page = () => {
-  return (
-    <IfUserThenIfAdmin>
-      <PageMergin>
-        <ProductForm />
-      </PageMergin>
-    </IfUserThenIfAdmin>
-  );
-};
-
-export default Page;
