--- client_admin_panel/src/app/(admin-routes)/(product-t1)/product-t1-edit-form/layout.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Layout = ({ children }: { children: React.ReactNode }) => {
-  return <>{children}</>;
-};
-
-export default Layout;
