+++ client_admin_panel/src/app/(admin-routes)/(product-t1)/product-t1-product-list/page.tsx
@@ -1,13 +1,13 @@
 import React from "react";
 import IfUserThenIfAdmin from "@/providers/security/if-user-then-if-admin";
 import PageMergin from "@/wrapper/page-mergin";
-import CustomTabT1 from "@/components/product-t1/products/ui-change-tab/custom-tab-t1";
+import UiTabChange from "@/components/product-t1/products/list/ui-tab-change";
 
 const Page = () => {
   return (
     <IfUserThenIfAdmin>
       <PageMergin>
-        <CustomTabT1 />
+        <UiTabChange />
       </PageMergin>
     </IfUserThenIfAdmin>
   );
