+++ client_admin_panel/src/app/(admin-routes)/(product-t1)/product-t1-update-banner/page.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import UpdateBanner from "@/components/product-t1/update/update-banner";
+import UpdateBanner from "@/components/product-t1/operation/banner/update-banner";
 import IfUserThenIfAdmin from "@/providers/security/if-user-then-if-admin";
 import PageMergin from "@/wrapper/page-mergin";
