--- client_admin_panel/src/app/@modal/(.)product-t1-edit-form/[id]/page.tsx
@@ -1,13 +1,0 @@
-import React from "react";
-import ProductForm from "@/components/product-t1/update/update-product";
-import PageMergin from "@/wrapper/page-mergin";
-
-const Page = () => {
-  return (
-    <PageMergin>
-      <ProductForm />
-    </PageMergin>
-  );
-};
-
-export default Page;
