+++ client_admin_panel/src/app/@modal/(.)product-t1-update-form/[id]/page.tsx
@@ -1,0 +1,19 @@
+"use client";
+
+import IfUserThenIfAdmin from "@/providers/security/if-user-then-if-admin";
+import PageMergin from "@/wrapper/page-mergin";
+import ActionContent from "@/components/product-t1/grid/admin/action/action-content";
+
+const Page = () => {
+  return (
+    <>
+      <IfUserThenIfAdmin>
+        <PageMergin>
+          <ActionContent />
+        </PageMergin>
+      </IfUserThenIfAdmin>
+    </>
+  );
+};
+
+export default Page;
