+++ client_admin_panel/src/components/product-t1/core/pagination/pagination.tsx
@@ -11,7 +11,6 @@ export function PaginationPage({
   totalItems: number;
 }) {
   const totalPages = Math.ceil(totalItems / ITEMS_PER_PAGE);
-
   return (
     <div className="flex items-center justify-between border-t border-gray-200 px-4 py-3 sm:px-6">
       <div className="hidden lg:block">
