--- client_admin_panel/src/components/product-t1/grid/admin/admin-info/card-admin-info.tsx
@@ -1,17 +1,0 @@
-import React from "react";
-import { Badge } from "@/components/ui/shadcn/badge";
-
-const CardAdminInfo = ({ product }: { product: any }) => {
-  return (
-    <div className="flex gap-2 mx-2 mb-2">
-      {product?.deleted && <Badge variant="destructive">Deleted</Badge>}
-      {product.stock > 0 ? (
-        <Badge>stock: {Math.min(product.stock ?? 0, 100000000)}</Badge>
-      ) : (
-        <Badge variant="destructive">Out of stock</Badge>
-      )}
-    </div>
-  );
-};
-
-export default CardAdminInfo;
