+++ client_admin_panel/src/components/product-t1/grid/admin/cards/list-view-cards.tsx
@@ -1,15 +1,12 @@
 import React from "react";
 import { CustomCardT2 } from "@/components/product-t1/grid/common/carousel/custom-card-t2";
 import ListViewCardsInfo from "@/components/product-t1/grid/common/card-info/list-view-cards-info";
-import CardAdminInfo from "@/components/product-t1/grid/admin/admin-info/card-admin-info";
+import CardInfo from "@/components/product-t1/grid/admin/info/card-info";
 import { ListViewController } from "@/components/product-t1/grid/grid-control";
 import { Card } from "@/components/ui/shadcn/card";
-import {
-  editProduct,
-  OPEN_MODAL,
-  type1ProductDetails,
-} from "@/links/product-list";
-import AdminActionUpdate from "@/components/product-t1/grid/admin/admin-action/admin-action-update";
+import { OPEN_MODAL, type1ProductDetails } from "@/links/product-list";
+import ActionUpdate from "@/components/product-t1/grid/admin/action/action-update";
+import { editProduct } from "@/links/product-update";
 
 const ListViewCards = ({ products }: { products: any }) => {
   return (
@@ -32,9 +29,9 @@ const ListViewCards = ({ products }: { products: any }) => {
             product={product}
             href={type1ProductDetails + product.id + OPEN_MODAL}
           />
-          <CardAdminInfo product={product} />
-          <AdminActionUpdate
-            href={editProduct}
+          <CardInfo product={product} />
+          <ActionUpdate
+            href={editProduct + product.id}
             className="absolute top-0 right-0 z-10 mt-2"
           />
         </Card>
