+++ client_admin_panel/src/components/product-t1/operation/banner/update-banner.tsx
