+++ client_admin_panel/src/components/product-t1/operation/product/preview/product-preview.tsx
