+++ client_admin_panel/src/components/product-t1/operation/product/update/no-url-parameter/validation.ts
@@ -1,0 +1,3 @@
+export const isValidObjectId = (id: string): boolean => {
+  return /^[0-9a-fA-F]{24}$/.test(id);
+};
