+++ client_admin_panel/src/components/product-t1/products/details/details.tsx
