+++ client_admin_panel/src/components/product-t1/products/list/user/product-main.tsx
