--- client_admin_panel/src/components/product-t1/products/ui-change-tab/custom-tab-t1.tsx
@@ -1,28 +1,0 @@
-import React from "react";
-import {
-  Tabs,
-  TabsContent,
-  TabsList,
-  TabsTrigger,
-} from "@/components/ui/shadcn/tabs";
-import { AdminPcComponentProductList } from "@/components/product-t1/products/admin-product-main/product-main";
-import { UserPcComponentProductList } from "@/components/product-t1/products/user-product-main/product-main";
-
-const CustomTabT1 = () => {
-  return (
-    <Tabs defaultValue="admin">
-      <TabsList className="default-card">
-        <TabsTrigger value="admin">Admin View</TabsTrigger>
-        <TabsTrigger value="user">User View</TabsTrigger>
-      </TabsList>
-      <TabsContent value="admin">
-        <AdminPcComponentProductList />
-      </TabsContent>
-      <TabsContent value="user">
-        <UserPcComponentProductList />
-      </TabsContent>
-    </Tabs>
-  );
-};
-
-export default CustomTabT1;
