+++ client_admin_panel/src/links/product-list.ts
@@ -13,5 +13,6 @@ export const type1ProductDetails = "/product-t1-details/";
 
 //Use: Product List Grid
 //PURPOSE: Use for redirect edit product page when user clicks on the edit button in the product card in the product list grid.
+//Example: /product-t1-update-form/1/?showDialog=y -> /product-t1-update-form/{productId}/?showDialog=y
 
-export const editProduct = "#";
+export const addProduct = "/product-t1-add-form";
