+++ client_admin_panel/src/links/product-update.ts
@@ -1,0 +1,25 @@
+//Use: Flags for modals
+//PURPOSE: If you pass this flag modal will be open.
+export const OPEN_MODAL = "/?showDialog=y";
+
+// ()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()
+
+//Use: Update Product/No URL Parameter
+//PURPOSE: If no id is found in the URL or the id is not valid, then redirect to this page.
+//Example: /product-t1-update/ -> /product-t1-update/no-url-parameter
+export const REDIRECT_TO_THE_PRODUCT_LIST = "/product-t1-product-list";
+
+// ()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()
+
+//Use: Update Product/No URL Parameter
+//PURPOSE: If no id is found in the URL or the id is not valid, then redirect to this page.
+//Example: /product-t1-update/ -> /product-t1-update/no-url-parameter
+export const editProduct = "/product-t1-update-form/";
+
+// ()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()()
+
+//Use: Update Product/No URL Parameter
+//PURPOSE: If no id is found in the URL or the id is not valid, then redirect to this page.
+//Example: /product-t1-update/ -> /product-t1-update/no-url-parameter
+export const REDIRECT_TO_THE_PRODUCT_UPDATE_UPDATE =
+  "/product-t1-update-form/update/";
